+++ inc/tickettemplatehiddenfield.class.php
@@ -42,5 +42,4 @@ class TicketTemplateHiddenField extends ITILTemplateHiddenField {
    static public $itemtype  = 'TicketTemplate';
    static public $items_id  = 'tickettemplates_id';
    static public $itiltype = 'Ticket';
-
 }